+++ delivery/models.py
@@ -1,45 +1,10 @@
-from flask import jsonify, request
-from flask.views import MethodView
 from dataclasses import dataclass
-from marshmallow import ValidationError
-from delivery.repositories import AbstractRestaurantRepository, MemoryRestaurantRepository
 
 
 @dataclass
 class Restaurant:
-    name: str
-    address: str
-    work_time: str
-    phone_number: str
+    name: str = ""
+    address: str = ""
+    work_time: str = ""
+    phone_number: str = ""
     id: int = 0
-
-
-class RestaurantEndpoint(MethodView):
-    def __init__(self, restaurants: AbstractRestaurantRepository, schema):
-        self.restaurants = restaurants
-        self.schema = schema
-
-    def get(self):
-        return jsonify([self.schema.dump(restaurant) for restaurant in self.restaurants.get_all()]), 200
-
-    def post(self):
-        content = request.json
-        self.restaurants.create(self.schema.load(content))
-        return jsonify([self.schema.dump(restaurant) for restaurant in self.restaurants.get_all()]), 201
-
-
-class RestaurantItemEndpoint(MethodView):
-    def __init__(self, restaurants: AbstractRestaurantRepository, schema):
-        self.restaurants = restaurants
-        self.schema = schema
-
-    def get(self, restaurant_id: int):
-        return jsonify(self.schema.dump(self.restaurants.get_by_id(restaurant_id))), 200
-
-    def put(self, restaurant_id: int):
-        content = request.json
-        errors = self.schema.validate(content)
-        if errors != {}:
-            raise ValidationError(errors)
-        self.restaurants.update(content, restaurant_id)
-        return jsonify(self.schema.dump(self.restaurants.get_by_id(restaurant_id))), 200
